+++ lib/heroku/api/postgres/backups.rb
@@ -23,22 +23,19 @@ module Heroku
           @client.perform_get_request("/client/v11/apps/#{app_id}/transfers/#{backup_id}")
         end
 
-        def schedules(app_id, database_id)
-          @client.perform_get_request("/client/v11/databases/#{database_id}/transfer-schedules",
-                                      host: db_host(app_id, database_id))
+        def schedules(database_id)
+          @client.perform_get_request("/client/v11/databases/#{database_id}/transfer-schedules")
         end
 
-        def schedule(app_id, database_id, hour: 0o0, timezone: 'UTC')
+        def schedule(database_id, hour: 0, timezone: 'UTC')
           @client.perform_post_request("/client/v11/databases/#{database_id}/transfer-schedules",
                                        { hour: hour,
                                          timezone: timezone,
-                                         schedule_name: 'DATABASE_URL' }, host: db_host(app_id, database_id))
+                                         schedule_name: 'DATABASE_URL' })
         end
 
-        def capture(app_id, database_id, options = {})
-          @client.perform_post_request("/client/v11/databases/#{database_id}/backups",
-                                       options,
-                                       host: db_host(app_id, database_id))
+        def capture(database_id, options = {})
+          @client.perform_post_request("/client/v11/databases/#{database_id}/backups", options)
         end
 
         def url(app_id, backup_num = nil)
@@ -64,15 +61,9 @@ module Heroku
           backup
         end
 
-        def restore(app_id, database_id, backup_url)
+        def restore(database_id, backup_url)
           @client.perform_post_request("/client/v11/databases/#{database_id}/restores",
-                                       { backup_url: backup_url }, host: db_host(app_id, database_id))
-        end
-
-        private
-
-        def db_host(app_id, database_id)
-          @client.db_host(app_id, database_id)
+                                       { backup_url: backup_url })
         end
       end
     end