+++ lib/heroku/api/postgres/credentials.rb
@@ -17,16 +17,11 @@ module Heroku
         # @param app_id [String] the application name.
         # @param database_id [String] the database UUID.
         # @param name [String] the credential name to be rotated (default: `default`).
-        def rotate(app_id, database_id, name: 'default')
+        def rotate(database_id, name: 'default')
           path = "/postgres/v0/databases/#{database_id}/credentials" \
                  "/#{URI.encode_www_form_component(name)}/credentials_rotation"
-          @client.perform_post_request(path, {}, host: db_host(app_id, database_id))
-        end
 
-        private
-
-        def db_host(app_id, database_id)
-          @client.db_host(app_id, database_id)
+          @client.perform_post_request(path, {})
         end
       end
     end