+++ lib/heroku/api/postgres/databases.rb
@@ -13,10 +13,10 @@ module Heroku
 
         # original call returns simply a database object, therefore I call the info API.
         # perform_get_request "/client/v11/databases/#{database_id}/wait_status"
-        def wait(app_id, database_id, options = { wait_interval: 3 })
+        def wait(database_id, options = { wait_interval: 3 })
           waiting = true
           while waiting
-            database = info(app_id, database_id)
+            database = info(database_id)
             break unless database[:waiting?]
 
             sleep(options[:wait_interval])
@@ -24,14 +24,8 @@ module Heroku
           database
         end
 
-        def info(app_id, database_id)
-          @client.perform_get_request("/client/v11/databases/#{database_id}", host: db_host(app_id, database_id))
-        end
-
-        private
-
-        def db_host(app_id, database_id)
-          @client.db_host(app_id, database_id)
+        def info(database_id)
+          @client.perform_get_request("/client/v11/databases/#{database_id}")
         end
       end
     end