+++ spec/heroku/api/postgres/client_spec.rb
@@ -8,13 +8,13 @@ RSpec.describe Heroku::Api::Postgres::Client, :vcr do
     let(:oauth_token) { ENV.fetch('HEROKU_OAUTH_TOKEN', nil) }
     it {
       expect do
-        Heroku::Api::Postgres.connect_oauth(oauth_token).databases.info(app_id, database_id)
+        Heroku::Api::Postgres.connect_oauth(oauth_token).databases.info(database_id)
       end.not_to raise_error
     }
   end
 
   describe '.connect' do
     let(:api_key) { ENV.fetch('HEROKU_API_KEY', nil) }
-    it { expect { Heroku::Api::Postgres.connect(api_key).databases.info(app_id, database_id) }.not_to raise_error }
+    it { expect { Heroku::Api::Postgres.connect(api_key).databases.info(database_id) }.not_to raise_error }
   end
 end